+++ src/bart/bart_model.rs
@@ -1041,6 +1041,9 @@ impl PrivateLanguageGenerator for BartGenerator {
     fn _get_tokenizer(&self) -> &TokenizerOption {
         &self.tokenizer
     }
+    fn _get_tokenizer_mut(&mut self) -> &mut TokenizerOption {
+        &mut self.tokenizer
+    }
     fn get_var_store(&self) -> &nn::VarStore {
         &self.var_store
     }