+++ src/pipelines/question_answering.rs
@@ -684,6 +684,16 @@ impl QuestionAnsweringModel {
         })
     }
 
+    /// Get a reference to the model tokenizer.
+    pub fn get_tokenizer(&self) -> &TokenizerOption {
+        &self.tokenizer
+    }
+
+    /// Get a mutable reference to the model tokenizer.
+    pub fn get_tokenizer_mut(&mut self) -> &mut TokenizerOption {
+        &mut self.tokenizer
+    }
+
     /// Perform extractive question answering given a list of `QaInputs`
     ///
     /// # Arguments