+++ src/pipelines/sequence_classification.rs
@@ -637,6 +637,16 @@ impl SequenceClassificationModel {
         })
     }
 
+    /// Get a reference to the model tokenizer.
+    pub fn get_tokenizer(&self) -> &TokenizerOption {
+        &self.tokenizer
+    }
+
+    /// Get a mutable reference to the model tokenizer.
+    pub fn get_tokenizer_mut(&mut self) -> &mut TokenizerOption {
+        &mut self.tokenizer
+    }
+
     fn prepare_for_model<'a, S>(&self, input: S) -> Tensor
     where
         S: AsRef<[&'a str]>,